Move function interpretation to separate files
