+++ ksmt-z3/src/main/kotlin/org/ksmt/solver/z3/KZ3Model.kt
@@ -13,6 +13,10 @@ import org.ksmt.KContext
 import org.ksmt.decl.KDecl
 import org.ksmt.expr.KExpr
 import org.ksmt.expr.KUninterpretedSortValue
+import org.ksmt.solver.model.KFuncInterp
+import org.ksmt.solver.model.KFuncInterpEntryWithVars
+import org.ksmt.solver.model.KFuncInterpVarsFree
+import org.ksmt.solver.model.KFuncInterpWithVars
 import org.ksmt.solver.KModel
 import org.ksmt.solver.model.KModelImpl
 import org.ksmt.sort.KSort
@@ -51,7 +55,7 @@ open class KZ3Model(
         constantDeclarations + functionDeclarations
     }
 
-    private val interpretations = hashMapOf<KDecl<*>, KModel.KFuncInterp<*>?>()
+    private val interpretations = hashMapOf<KDecl<*>, KFuncInterp<*>?>()
     private val uninterpretedSortValues = hashMapOf<KUninterpretedSort, UninterpretedSortValueContext>()
 
     private fun loadConstDeclarations(): Set<KDecl<KSort>> {
@@ -82,7 +86,7 @@ open class KZ3Model(
         return with(converter) { z3Result.convertExpr() }
     }
 
-    override fun <T : KSort> interpretation(decl: KDecl<T>): KModel.KFuncInterp<T>? =
+    override fun <T : KSort> interpretation(decl: KDecl<T>): KFuncInterp<T>? =
         interpretations.getOrPut(decl) {
             ctx.ensureContextMatch(decl)
             ensureContextActive()
@@ -104,13 +108,13 @@ open class KZ3Model(
     internal fun resolveUninterpretedSortValue(sort: KUninterpretedSort, decl: Long): KUninterpretedSortValue =
         getUninterpretedSortContext(sort).getValue(decl)
 
-    private fun <T : KSort> constInterp(decl: KDecl<T>, z3Decl: Long): KModel.KFuncInterp<T>? {
+    private fun <T : KSort> constInterp(decl: KDecl<T>, z3Decl: Long): KFuncInterp<T>? {
         val z3Interp = model.getConstInterp(z3Decl) ?: return null
         val expr = with(converter) { z3Interp.convertExpr<T>() }
-        return KModel.KFuncInterpVarsFree(decl = decl, entries = emptyList(), default = expr)
+        return KFuncInterpVarsFree(decl = decl, entries = emptyList(), default = expr)
     }
 
-    private fun <T : KSort> funcInterp(decl: KDecl<T>, z3Decl: Long): KModel.KFuncInterp<T>? = with(converter) {
+    private fun <T : KSort> funcInterp(decl: KDecl<T>, z3Decl: Long): KFuncInterp<T>? = with(converter) {
         val z3Interp = model.getFuncInterp(z3Decl) ?: return null
 
         val vars = decl.argSorts.map { it.mkFreshConst("x") }
@@ -121,13 +125,13 @@ open class KZ3Model(
         val entries = z3Interp.entries.map { entry ->
             val args = entry.args.map { it.substituteVarsAndConvert<KSort>(z3Vars) }
             val value = entry.value.substituteVarsAndConvert<T>(z3Vars)
-            KModel.KFuncInterpEntryWithVars.create(args, value)
+            KFuncInterpEntryWithVars.create(args, value)
         }
 
         val default = z3Interp.elseExpr.substituteVarsAndConvert<T>(z3Vars)
         val varDecls = vars.map { it.decl }
 
-        return KModel.KFuncInterpWithVars(decl, varDecls, entries, default)
+        return KFuncInterpWithVars(decl, varDecls, entries, default)
     }
 
     override fun detach(): KModel {
